Mistake in sorting of pairs

+++ src/libgcm/util/Types.hpp
@@ -291,7 +291,7 @@ namespace gcm
 
     // For pair sorting
     struct sort_pred {
-        bool operator()(const std::pair<int,int> &left, const std::pair<int,int> &right) {
+        bool operator()(const std::pair<int,float> &left, const std::pair<int,float> &right) {
             return left.second < right.second;
         }
     };
